simulators/robot/requirements.txt
google-cloud-pubsub==2.21.6